04-BasicToAdvanced-RAG/requirements.txt
aiofiles==23.2.1
aiohappyeyeballs==2.4.0
aiohttp==3.10.5
aiosignal==1.3.1
altair==5.4.1
annotated-types==0.7.0
annoy==1.17.3
anyio==4.4.0
attrs==24.2.0
audioread==3.0.1
beautifulsoup4==4.12.3
blinker==1.8.2
cachetools==5.5.0
certifi==2024.8.30
cffi==1.17.1
charset-normalizer==3.3.2
click==8.1.7
contourpy==1.3.0
cryptography==43.0.1
cycler==0.12.1
dataclasses-json==0.6.7
decorator==5.1.1
faiss-cpu==1.8.0.post1
fastapi==0.114.2
ffmpy==0.4.0
filelock==3.16.0
fonttools==4.53.1
frozenlist==1.4.1
fsspec==2024.9.0
gensim==4.3.3
gitdb==4.0.11
GitPython==3.1.43
gradio==4.44.0
gradio_client==1.3.0
h11==0.14.0
httpcore==1.0.5
httpx==0.27.2
huggingface-hub==0.24.7
idna==3.10
importlib_resources==6.4.5
Jinja2==3.1.4
joblib==1.4.2
jsonpatch==1.33
jsonpointer==3.0.0
jsonschema==4.23.0
jsonschema-specifications==2023.12.1
kiwisolver==1.4.7
langchain==0.3.0
langchain-community==0.3.0
langchain-core==0.3.0
langchain-text-splitters==0.3.0
langsmith==0.1.121
lazy_loader==0.4
librosa==0.10.2.post1
llvmlite==0.43.0
markdown-it-py==3.0.0
MarkupSafe==2.1.5
marshmallow==3.22.0
matplotlib==3.9.2
mdurl==0.1.2
mpmath==1.3.0
msgpack==1.1.0
multidict==6.1.0
mypy-extensions==1.0.0
narwhals==1.8.1
networkx==3.3
node2vec==0.5.0
numba==0.60.0
numpy==1.26.4
openai==0.28.0
orjson==3.10.7
packaging==24.1
pandas==2.2.2
pdfminer.six==20231228
pdfplumber==0.11.4
pillow==10.4.0
pinecone==5.1.0
pinecone-plugin-inference==1.1.0
pinecone-plugin-interface==0.0.7
platformdirs==4.3.3
pooch==1.8.2
protobuf==5.28.1
pyarrow==17.0.0
pycparser==2.22
pydantic==2.9.1
pydantic-settings==2.5.2
pydantic_core==2.23.3
pydeck==0.9.1
pydub==0.25.1
Pygments==2.18.0
pyparsing==3.1.4
PyPDF2==3.0.1
pypdfium2==4.30.0
python-dateutil==2.9.0.post0
python-dotenv==1.0.1
python-multipart==0.0.9
pytz==2024.2
PyYAML==6.0.2
referencing==0.35.1
regex==2024.9.11
requests==2.32.3
rich==13.8.1
rpds-py==0.20.0
ruff==0.6.5
safetensors==0.4.5
scikit-learn==1.5.2
scipy==1.13.1
semantic-version==2.10.0
sentence-transformers==3.1.0
shellingham==1.5.4
six==1.16.0
smart-open==7.0.4
smmap==5.0.1
sniffio==1.3.1
soundfile==0.12.1
soupsieve==2.6
soxr==0.5.0.post1
SQLAlchemy==2.0.35
starlette==0.38.5
streamlit==1.38.0
sympy==1.13.2
tenacity==8.5.0
threadpoolctl==3.5.0
tokenizers==0.19.1
toml==0.10.2
tomlkit==0.12.0
torch==2.4.1
torchvision==0.19.1
tornado==6.4.1
tqdm==4.66.5
transformers==4.44.2
typer==0.12.5
typing-inspect==0.9.0
typing_extensions==4.12.2
tzdata==2024.1
urllib3==2.2.3
uvicorn==0.30.6
websockets==12.0
wrapt==1.16.0
yarl==1.11.1
